+++ lib/route_animation_helper.dart
@@ -7,14 +7,15 @@ import 'package:flutter/material.dart';
 import 'dart:math';
 
 /// enum representing the selected animation type you want to apply
-/// 
+///
 /// [AnimType.slideStart] - enter the next screen sliding from the side of the screen.
 /// [AnimType.slideBottom] - enter the next screen from the bottom of the screen.
-/// [AnimType.scale] - opens the next screen from the middle of the screen and expanding to the top and bottom of the screen.
+/// [AnimType.size] - opens the next screen from the middle of the screen and expanding to the top and bottom of the screen.
+/// [AnimType.scale] - scales the next screen from the center till full screen.
 /// [AnimType.fade] - opens the next screen by fading it in.
 /// [AnimType.rotate] - rotates the next screen 360 degrees.
 /// [AnimType.cubic] - 3D cubic animation between screens.
-enum AnimType { slideStart, slideBottom, scale, fade, rotate, cubic }
+enum AnimType { slideStart, slideBottom, scale, size, fade, rotate, cubic }
 
 /// Helper class to generate an Animated Route for Screens transition
 ///
@@ -33,10 +34,12 @@ class RouteAnimationHelper {
   static const int DEFAULT_TRANSITION_DURATION = 450;
 
   static Route createRoute(
-      {Widget currentPage, Widget destination,
+      {Widget currentPage,
+      Widget destination,
       Curve curve = Curves.ease,
       AnimType animType = AnimType.slideStart,
-      int duration = DEFAULT_TRANSITION_DURATION, Color cubicBackgroundColor = Colors.white}) {
+      int duration = DEFAULT_TRANSITION_DURATION,
+      Color cubicBackgroundColor = Colors.white}) {
     return PageRouteBuilder(
         pageBuilder: (context, animation, anotherAnimation) {
           return animType == AnimType.cubic ? currentPage : destination;
@@ -44,7 +47,7 @@ class RouteAnimationHelper {
         transitionDuration: Duration(milliseconds: duration),
         transitionsBuilder: (context, animation, anotherAnimation, child) {
           animation = CurvedAnimation(curve: curve, parent: animation);
-          
+
           switch (animType) {
             case AnimType.fade:
               return Align(
@@ -67,12 +70,19 @@ class RouteAnimationHelper {
                 child: child,
               );
               break;
-            case AnimType.scale:
+            case AnimType.size:
               return Align(
                 child: SizeTransition(
                   sizeFactor: animation,
                   child: child,
                   axisAlignment: 0.0,
+                ),
+              );
+            case AnimType.scale:
+              return Align(
+                child: ScaleTransition(
+                  scale: animation,
+                  child: child,
                 ),
               );
               break;
@@ -82,16 +92,17 @@ class RouteAnimationHelper {
                 child: child,
               );
             case AnimType.cubic:
-              
-              if(currentPage == null){
-                print('RouteAnimationHelper, currentPage param is mandatory for cubic transition...');
+              if (currentPage == null) {
+                print(
+                    'RouteAnimationHelper, currentPage param is mandatory for cubic transition...');
                 return SlideTransition(
-                  position: Tween(begin: Offset(1.0, 0.0), end: Offset(0.0, 0.0))
-                      .animate(animation),
+                  position:
+                      Tween(begin: Offset(1.0, 0.0), end: Offset(0.0, 0.0))
+                          .animate(animation),
                   child: child,
-                ); 
+                );
               }
-              
+
               return Stack(
                 children: <Widget>[
                   SlideTransition(